Fix: Beta fixes for types, remove inline props
