+++ src/components/Active/Active.ts
@@ -6,7 +6,7 @@ import {
   type ActiveContainerBaseProps,
   type ActiveContentBaseProps,
   type ComponentBaseProps,
-} from '../../utils/types'
+} from '../../utils/base-types'
 import { Link } from '../Link/Link'
 
 export interface ActiveProps
