+++ src/components/Alert/Alert.tsx
@@ -2,7 +2,8 @@ import React from 'react'
 import { closeBase } from '../Close/Close'
 import { useTheme } from '../Theme/Theme'
 import { useActive, useVisible } from '../../hooks'
-import { type ComponentBaseProps, type MergePropTypes } from '../../utils/types'
+import { type ComponentBaseProps } from '../../utils/base-types'
+import { type MergePropTypes } from '../../utils/types'
 import { element } from '../../utils/element-creator'
 import { staticComponent } from '../../utils/static-component-builder'
