+++ src/components/Badge/Badge.ts
@@ -1,6 +1,6 @@
 import { useTheme } from '../Theme/Theme'
 import { element } from '../../utils/element-creator'
-import { type ComponentBaseProps } from '../../utils/types'
+import { type ComponentBaseProps } from '../../utils/base-types'
 
 interface BadgeProps extends ComponentBaseProps<'div'> {
   /** Variant color */
