+++ src/components/Block/Block.ts
@@ -1,28 +1,19 @@
 import React from 'react'
 import { useTheme } from '../Theme/Theme'
-import { type ComponentBaseProps } from '../../utils/types'
+import { type ComponentBaseProps } from '../../utils/base-types'
 import { element } from '../../utils/element-creator'
 
-export interface BlockProps extends ComponentBaseProps<'div'> {
-  /** Sets display between to an inline or block element */
-  inline?: boolean
-}
+export interface BlockProps extends ComponentBaseProps<'div'> {}
 
 /**
  * [Block component 📝](https://componentry.design/components/block)
  */
 export const Block: React.FC<BlockProps> = (props) => {
-  const { inline, ...rest } = {
+  const rest = {
     ...useTheme<BlockProps>('Block'),
     ...props,
   }
 
-  return element({
-    componentCx: {
-      'block': !inline,
-      'inline-block': inline,
-    },
-    ...rest,
-  })
+  return element(rest)
 }
 Block.displayName = 'Block'
