+++ src/components/Flex/Flex.ts
@@ -1,6 +1,6 @@
 import React from 'react'
 import { useTheme } from '../Theme/Theme'
-import { type ComponentBaseProps } from '../../utils/types'
+import { type ComponentBaseProps } from '../../utils/base-types'
 import { element } from '../../utils/element-creator'
 
 export interface FlexProps extends ComponentBaseProps<'div'> {
@@ -8,8 +8,6 @@ export interface FlexProps extends ComponentBaseProps<'div'> {
   align?: 'start' | 'end' | 'center' | 'baseline' | 'stretch'
   /** Sets a `flex-direction` flex style */
   direction?: 'column' | 'column-reverse' | 'row-reverse' | 'row'
-  /** Switches between display between an inline and block element */
-  inline?: boolean
   /** Sets a `justify-content` style */
   justify?: 'start' | 'end' | 'center' | 'between' | 'around' | 'evenly'
   /** Sets a `flex-wrap` flex style */
@@ -20,22 +18,16 @@ export interface FlexProps extends ComponentBaseProps<'div'> {
  * [Flex component 📝](https://componentry.design/components/flex)
  */
 export const Flex: React.FC<FlexProps> = (props) => {
-  const { align, direction, inline, justify, wrap, ...rest } = {
+  const { align, direction, justify, wrap, ...rest } = {
     ...useTheme<FlexProps>('Flex'),
     ...props,
   }
-  // Tailwind uses a flex-col class but direction="col" is super wonky
-  // => so props use "column" and we replace with "col"
-  const computedDirection = direction?.replace('column', 'col')
 
   return element({
-    componentCx: {
-      'flex': !inline,
-      'inline-flex': inline,
-      [`flex-${computedDirection}`]: computedDirection,
-      [`flex-${wrap}`]: wrap,
-    },
+    display: 'flex',
     alignItems: align,
+    flexDirection: direction,
+    flexWrap: wrap,
     justifyContent: justify,
     ...rest,
   })
