+++ src/components/Grid/Grid.stories.mdx
@@ -31,19 +31,4 @@ export function padNum(num) {
 Grid is a layout component with shorthand properties `align`, and `justify` for common
 alignment styles.
 
-## Inline block elements
-
-Create inline grid containers using the `inline` prop:
-
-<Grid inline gap={2} backgroundColor='primary-200' className='rounded grid-cols-2'>
-  {[1, 2].map((key) => (
-    <div
-      key={key}
-      className='bg-primary-600 text-inverse font-mono font-bold px-3 py-2 rounded'
-    >
-      #{padNum(key)}
-    </div>
-  ))}
-</Grid>
-
 <ArgsTable of={Grid} />
