+++ src/components/Text/Text.stories.mdx
@@ -47,20 +47,6 @@ configured using the `& + &` stle in the body variant styles.
   your dinosaur tour, right?
 </Text>
 
-## Inline text elements
-
-<Text>
-  Styling inline text elements is easiest with a Block component and the{' '}
-  <Block inline bold>
-    inline
-  </Block>{' '}
-  prop. This creates an inline div with access to all{' '}
-  <Block inline italic>
-    utility
-  </Block>{' '}
-  props.
-</Text>
-
 ## Utility styles - fontSize
 
 The `fontSize` prop is intended to be used with the `body` variant:
