+++ src/components/Text/Text.ts
@@ -1,6 +1,7 @@
 import React from 'react'
 import { useTheme } from '../Theme/Theme'
-import { type ComponentBaseProps, type MergePropTypes } from '../../utils/types'
+import { type ComponentBaseProps } from '../../utils/base-types'
+import { type MergePropTypes } from '../../utils/types'
 import { element } from '../../utils/element-creator'
 
 /** Element used for each variant */
@@ -21,8 +22,6 @@ export interface TextPropsOverrides {}
 interface TextPropsDefaults {
   /** Display variant */
   variant?: 'h1' | 'h2' | 'h3' | 'body' | 'code' | 'small'
-  /** Sets a bold font-weight style */
-  bold?: boolean
 }
 
 type TextProps = MergePropTypes<TextPropsDefaults, TextPropsOverrides> &
@@ -35,7 +34,6 @@ export const Text: React.FC<TextProps> = (props) => {
   const {
     variant = 'body',
     elementsMap = defaultElementsMap,
-    bold,
     ...rest
   } = {
     ...useTheme<TextProps & { elementsMap?: ElementsMap }>('Text'),
@@ -45,7 +43,6 @@ export const Text: React.FC<TextProps> = (props) => {
   return element({
     as: elementsMap[variant],
     componentCx: `🅲Text-base 🅲Text-${variant}`,
-    fontWeight: bold ? 'bold' : undefined,
     ...rest,
   })
 }
