+++ src/components/Tooltip/Tooltip.tsx
@@ -8,7 +8,7 @@ import {
   type ActiveContainerBaseProps,
   type ActiveContentBaseProps,
   type ComponentBaseProps,
-} from '../../utils/types'
+} from '../../utils/base-types'
 import { Link } from '../Link/Link'
 
 export interface TooltipProps
