+++ src/index.ts
@@ -29,4 +29,4 @@ export { Tooltip } from './components/Tooltip/Tooltip'
 export { useActive, useActiveScrollReset, useNoScroll, useVisible } from './hooks'
 export { setupOutlineHandlers } from './utils/dom'
 export { utilityClasses } from './utils/utility-classes'
-export { type UtilityProps } from './utils/types'
+export { type UtilityProps } from './utils/utility-props'
