+++ src/utils/active-action-component-builder.tsx
@@ -2,7 +2,7 @@ import React, { useContext } from 'react'
 import { useTheme } from '../components/Theme/Theme'
 import { ActiveCtx } from './active-container-component-builder'
 import { type ARIAControls, computeARIA } from './aria'
-import { type ActiveActionBaseProps } from './types'
+import { type ActiveActionBaseProps } from './base-types'
 import { element } from './element-creator'
 
 interface ActiveActionBuilder {
