+++ src/utils/utility-props.ts
@@ -1,0 +1,137 @@
+/**
+ * @file Defines the allowed values for shared Componentry utility props.
+ * @remarks
+ * This file is separate from `base-types` for easier reference and usage of
+ * module augmentation of utility props, eg:
+ *
+ * `declare module 'componentry/types/utils/utility-props' { }`
+ */
+
+import { MergePropTypes } from './types'
+
+/** Default utility prop types, customizable with UtilityPropsOverrides */
+export interface UtilityPropsDefaults {
+  /** Sets active style */
+  active?: boolean | string
+  /** Sets align-content style */
+  alignContent?: 'start' | 'end' | 'center' | 'between' | 'around' | 'stretch'
+  /** Sets align-items style */
+  alignItems?: 'start' | 'end' | 'center' | 'baseline' | 'stretch'
+  /** Sets align-self style */
+  alignSelf?: 'auto' | 'start' | 'end' | 'center' | 'baseline' | 'stretch'
+  /** Sets background color style */
+  backgroundColor?: 'primary' | 'success' | 'warning' | 'critical'
+  /** Sets a bold font weight style */
+  bold?: boolean
+  /** Sets a 1px border-width style */
+  border?: boolean
+  /** Sets a border-width size style */
+  borderWidth?: string | number
+  /** Sets a 1px border-bottom-width style */
+  borderBottom?: boolean
+  /** Sets border color style */
+  borderColor?: 'primary'
+  /** Sets a 1px border-left-width style */
+  borderLeft?: boolean
+  /** Sets a 1px border-right-width style */
+  borderRight?: boolean
+  /** Sets a 1px border-top-width style */
+  borderTop?: boolean
+  /** Sets a disabled style */
+  disabled?: boolean
+  /** Sets a display style */
+  display?:
+    | 'block'
+    | 'inline-block'
+    | 'inline'
+    | 'flex'
+    | 'inline-flex'
+    | 'flow-root'
+    | 'grid'
+    | 'inline-grid'
+    | 'contents'
+    | 'list-item'
+    | 'hidden'
+  /** Sets a text color style */
+  color?: string
+  /** Sets a `flex-direction` flex style */
+  flexDirection?: 'column' | 'column-reverse' | 'row-reverse' | 'row'
+  /** Sets a `flex-wrap` flex style */
+  flexWrap?: 'wrap' | 'nowrap' | 'wrap-reverse'
+  /** Sets a font-family style */
+  fontFamily?: 'body' | 'mono'
+  /** Sets a font-size style */
+  fontSize?: 'sm' | 'base' | 'lg'
+  /** Text font-weight style */
+  fontWeight?: 'light' | 'normal' | 'bold'
+  /** Sets height style */
+  height?: 'auto' | 'full' | 'screen' | 'min' | 'max' | 'fit' | string | number
+  /** Sets a display: none style */
+  invisible?: boolean
+  /** Sets an italic style */
+  italic?: boolean
+  /** Sets justify-content style */
+  justifyContent?: 'start' | 'end' | 'center' | 'between' | 'around' | 'evenly'
+  /** Sets a letter-spacing style */
+  letterSpacing?: 'tighter' | 'tight' | 'normal' | 'wide' | 'wider' | 'widest'
+  /** Sets a line-height style */
+  lineHeight?: 'none' | 'tight' | 'snug' | 'normal' | 'relaxed' | 'loose' | number
+  /** Sets a max-width style */
+  maxWidth?: 'none' | 'full' | 'min' | 'max' | 'fit' | 'prose' | 0 | string
+  /** Sets a max-height style */
+  maxHeight?: 'full' | 'screen' | 'min' | 'max' | 'fit' | number
+  /** Sets a min-height style */
+  minHeight?: 'full' | 'screen' | 'min' | 'max' | 'fit' | 0
+  /** Sets a min-width style */
+  minWidth?: 'full' | 'min' | 'max' | 'fit' | 0
+  /** Sets position style */
+  position?: 'static' | 'relative' | 'absolute' | 'fixed' | 'sticky'
+  /** Sets text-align style */
+  textAlign?: 'left' | 'center' | 'right' | 'justify'
+  /** Sets text-transform style */
+  textTransform?: 'lowercase' | 'uppercase' | 'capitalize' | 'normal-case'
+  /** Sets visible style */
+  visible?: boolean
+  /** Sets width style */
+  width?: 'auto' | 'full' | 'screen' | 'min' | 'max' | 'fit' | string | number
+
+  // --- Spacing
+  /** Sets a `gap` style */
+  gap?: string | number
+  /** margin */
+  m?: string | number
+  /** margin-top */
+  mt?: string | number
+  /** margin-right */
+  mr?: string | number
+  /** margin-bottom */
+  mb?: string | number
+  /** margin-left */
+  ml?: string | number
+  /** margin-left && margin-right */
+  mx?: string | number
+  /** margin-top && margin-bottom  */
+  my?: string | number
+  /** padding */
+  p?: string | number
+  /** padding-top */
+  pt?: string | number
+  /** padding-right */
+  pr?: string | number
+  /** padding-bottom */
+  pb?: string | number
+  /** padding-left */
+  pl?: string | number
+  /** padding-left && padding-right */
+  px?: string | number
+  /** padding-top && padding-bottom */
+  py?: string | number
+}
+
+/** Module augmentation interface for overriding default utility props' types */
+export interface UtilityPropsOverrides {}
+
+/**
+ * Componentry utility props for including utility styles.
+ */
+export type UtilityProps = MergePropTypes<UtilityPropsDefaults, UtilityPropsOverrides>
